seperated SearchHeader and SearchListView to other files
