+++ PriceTracker/PriceTracker/Views/Detail/DetailView.swift
