+++ PriceTracker/PriceTracker/Views/Search/NoContentView.swift
